// src/app/models/survey-filter.ts
export interface SurveyFilter {
  ageMax: number;
  ageMin: number;
  educations: string[];
  genders: string[];
  grossEarningsMax: number;
  grossEarningsMin: number;
  id: number;
  laborSectors: string[];
  maritalStatuses: string[];
  sizeOfTheHometownMax: number;
  sizeOfTheHometownMin: number;
  sizeOfTownMax: number;
  sizeOfTownMin: number;
  surveyId: number;
}